requirements.txt
duckdb
pandas
pyarrow
gdown
python-dotenv
streamlit
